"Refactor auth controller login logic and update avatar transformation in user controller"

+++ Back-End/controllers/auth.controller.js
@@ -107,6 +107,12 @@ exports.Login = asyncWrapper(async (req, res, next) => {
     );
   }
   const user = await User.findOne({ email }).select("+password");
+  if (!user || !(await user.correctPassword(password, user.password))) {
+    return next(AppError.create("Incorrect email or password", ERROR, 401));
+  }
+  if (!user.password) {
+    return next(AppError.create("Try login by another way", ERROR, 400));
+  }
   if (!user.isVerify) {
     await user.deleteOne();
     return next(
@@ -116,12 +122,6 @@ exports.Login = asyncWrapper(async (req, res, next) => {
         401
       )
     );
-  }
-  if (!user.password) {
-    return next(AppError.create("Try login by another way", ERROR, 400));
-  }
-  if (!user || !(await user.correctPassword(password, user.password))) {
-    return next(AppError.create("Incorrect email or password", ERROR, 401));
   }
   let token = await signToken(user,res);
   res.status(200).json({
